requirements.txt
pyvis
matplotlib
tqdm
groq
python-dotenv